pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "external-metadata-mappings"
version = "0.1.0"
description = "Schemas and APIs for PEP 725 External Metadata Mappings"
readme = "README.md"
authors = [
  { name = "Jaime Rodríguez-Guerra", email = "jrodriguez@quansight.com" },
]
license = { file = "LICENSE" }
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
requires-python = ">=3.10"
dependencies = [
  "requests",
  "jsonschema",
  "packageurl-python==0.16.0"
]

[project.urls]
homepage = "https://github.com/jaimergp/external-metadata-mappings"

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["linux-64", "osx-64", "osx-arm64", "win-64"]

[tool.pixi.dependencies]
packageurl-python = "==0.16.0"

[tool.pixi.pypi-dependencies]
external-metadata-mappings = { path = ".", editable = true }

[tool.pixi.feature.schemas.dependencies]
pydantic = ">=2"
[tool.pixi.feature.schemas.tasks]
schemas = "python schemas/schemas.py"

[tool.pixi.feature.streamlit.dependencies]
streamlit = "*"
[tool.pixi.feature.streamlit.tasks]
app = "python -m streamlit run --server.runOnSave=true app.py"

[tool.pixi.feature.pre-commit.dependencies]
pre-commit = "*"
[tool.pixi.feature.pre-commit.tasks]
pre-commit = "pre-commit"

[tool.pixi.feature.py310.dependencies]
python = "3.10.*"
[tool.pixi.feature.py311.dependencies]
python = "3.11.*"
[tool.pixi.feature.py312.dependencies]
python = "3.12.*"
[tool.pixi.feature.py313.dependencies]
python = "3.13.*"

[tool.pixi.feature.test.dependencies]
pytest = "*"

[tool.pixi.feature.test.tasks]
test = "pytest"

[tool.pixi.environments]
test-py310 = ["test", "py310"]
test-py311 = ["test", "py311"]
test-py312 = ["test", "py312"]
test-py313 = ["test", "py313"]
streamlit = ["streamlit"]
schemas = ["schemas"]
pre-commit = ["pre-commit"]
